Sloučení a úprava JS
