+++ symfony/assets/app.js
@@ -6,7 +6,6 @@
  */
 import './styles/app.css';
 import './script/script.js';
-import './script/toggle.js';
 
 console.log('This log comes from assets/app.js - welcome to AssetMapper! 🎉');
